+++ Tests/MapboxSearchTests/Common/Models/Search Result/ServerSearchResultTests.swift
@@ -3,35 +3,35 @@ import XCTest
 
 class ServerSearchResultTests: XCTestCase {
     func testServerSearchResultNilForUnknownType() {
-        let result = ServerSearchResult(coreResult: CoreSearchResultStub(id: UUID().uuidString, type: .unknown),
+        let result = ServerSearchResult(coreResult: CoreSearchResultStub(id: UUID().uuidString, mapboxId: "", type: .unknown),
                                         response: CoreSearchResponseStub.failureSample)
         
         XCTAssertNil(result)
     }
     
     func testServerSearchResultNilForCategoryType() {
-        let result = ServerSearchResult(coreResult: CoreSearchResultStub(id: UUID().uuidString, type: .category),
+        let result = ServerSearchResult(coreResult: CoreSearchResultStub(id: UUID().uuidString, mapboxId: "", type: .category),
                                         response: CoreSearchResponseStub.failureSample)
         
         XCTAssertNil(result)
     }
     
     func testServerSearchResultNilForUserRecordType() {
-        let result = ServerSearchResult(coreResult: CoreSearchResultStub(id: UUID().uuidString, type: .userRecord),
+        let result = ServerSearchResult(coreResult: CoreSearchResultStub(id: UUID().uuidString, mapboxId: "", type: .userRecord),
                                         response: CoreSearchResponseStub.failureSample)
         
         XCTAssertNil(result)
     }
     
     func testServerSearchResultNilForMissingCoordinates() {
-        let result = ServerSearchResult(coreResult: CoreSearchResultStub(id: UUID().uuidString, type: .userRecord, centerLocation: nil),
+        let result = ServerSearchResult(coreResult: CoreSearchResultStub(id: UUID().uuidString, mapboxId: "", type: .userRecord, centerLocation: nil),
                                         response: CoreSearchResponseStub.failureSample)
         
         XCTAssertNil(result)
     }
     
     func testServerSearchResultPOIFields() throws {
-        let coreResult = CoreSearchResultStub(id: UUID().uuidString, type: .poi)
+        let coreResult = CoreSearchResultStub(id: UUID().uuidString, mapboxId: "", type: .poi)
         let result = try XCTUnwrap(ServerSearchResult(coreResult: coreResult,
                                                       response: CoreSearchResponseStub.failureSample))
         XCTAssertEqual(result.suggestionType, .POI)
@@ -45,7 +45,7 @@ class ServerSearchResultTests: XCTestCase {
     }
     
     func testServerSearchResultAddressType() throws {
-        let coreResult = CoreSearchResultStub(id: UUID().uuidString, type: .address)
+        let coreResult = CoreSearchResultStub(id: UUID().uuidString, mapboxId: "", type: .address)
         let result = try XCTUnwrap(ServerSearchResult(coreResult: coreResult,
                                                       response: CoreSearchResponseStub.failureSample))
         XCTAssertEqual(result.suggestionType, .address(subtypes: [.address]))